+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.7 (unreleased)
 
+- Added metrics
 - Added `load` method
 
 ## 0.2.6 (2021-12-02)